修改自动任务，添加自动任务管理页面

+++ learn-shop-admin-user/src/main/java/com/billow/dao/TestDao.java
@@ -9,5 +9,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * @author liuyongtao
  * @create 2018-02-11 9:43
  */
-public interface TestDao extends JpaRepository<TestPo, Integer> {
+public interface TestDao extends JpaRepository<TestPo, Long> {
 }
