+++ learn-shop-core-order/src/main/java/com/billow/dao/OrderDao.java
@@ -3,5 +3,5 @@ package com.billow.dao;
 import com.billow.pojo.po.OrderPo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderDao extends JpaRepository<OrderPo, Integer> {
+public interface OrderDao extends JpaRepository<OrderPo, Long> {
 }
